Documentation

+++ src/d00_utils/data_retrieval.py
@@ -8,22 +8,40 @@ import re
 import shutil
 import datetime
 
-def retrieve_videos_from_s3(bDownload=True, from_date=None, to_date=None):
+def retrieve_videos_from_s3(bKeep_data=True, from_date=None, to_date=None):
+    """Retrieve jamcam videos from the s3 bucket based on the dates specified.
+    Downloads to a local directory and then loads them into numpy arrays.
 
+        Args:
+            bKeep_data: boolean for keeping the downloaded data in the local folder
+            from_date: start date (inclusive) for retrieving videos, if None then will retrieve from 2019-06-01 onwards
+            to_date: end date (inclusive) for retrieving vidoes, if None then will retrieve up to current day
+        Returns:
+            list of numpy arrays containing all the jamcam videos between the selected dates
+        Raises:
+
+    """
+
+    # Deal with the cases when a date isn't specified
     if (from_date == None):
         from_date = '2019-06-01'
     if (to_date == None):
         to_date = str(datetime.datetime.now())[:10]
 
-    s3_session = boto3.Session(profile_name='dssg')
-    bucket_name = 'air-pollution-uk'
+    # Set local directory for downloading data, will overwrite whatever is currently there
     local_dir = os.path.join(os.path.dirname(os.path.realpath(__file__)),
                              '..', '..', 'data/01_raw/jamcams/')
-    shutil.rmtree(local_dir)
+    if (os.path.isdir(local_dir)):
+        shutil.rmtree(local_dir)
     os.mkdir(local_dir)
+
+    # Set up boto3 session
+    s3_session = boto3.Session(profile_name='dssg')
+    bucket_name = 'air-pollution-uk'
     s3_resource = s3_session.resource('s3')
     my_bucket = s3_resource.Bucket(bucket_name)
 
+    # Get list of files in s3 based on dates provided
     selected_files = []
     objects = my_bucket.objects.filter(Prefix="raw/video_data_new/")
     files = [obj.key for obj in objects]
@@ -32,38 +50,31 @@ def retrieve_videos_from_s3(bDownload=True, from_date=None, to_date=None):
         if (date >= from_date and date <= to_date):
             selected_files.append(file)
 
+    # Download the selected files
     data = []
+    for file in selected_files:
+        my_bucket.download_file(file, local_dir + file.split('/')[-1])
 
-    if(bDownload):
+    # Load files into a list of numpy arrays using opencv
+    for file in os.listdir(local_dir):
+        cap = cv2.VideoCapture(local_dir + file)
 
-        for file in selected_files:
-            my_bucket.download_file(file, local_dir + file.split('/')[-1])
+        frameCount = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
+        frameWidth = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
+        frameHeight = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
 
-        for file in os.listdir(local_dir):
-            cap = cv2.VideoCapture(local_dir + file)
+        buf = np.empty((frameCount, frameHeight, frameWidth, 3), np.dtype('uint8'))
+        fc = 0
+        ret = True
+        while (fc < frameCount and ret):
+            ret, buf[fc] = cap.read()
+            fc += 1
+        cap.release()
+        data.append(buf)
 
-            frameCount = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
-            frameWidth = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
-            frameHeight = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
-
-            buf = np.empty((frameCount, frameHeight, frameWidth, 3), np.dtype('uint8'))
-            fc = 0
-            ret = True
-            while (fc < frameCount and ret):
-                ret, buf[fc] = cap.read()
-                fc += 1
-            cap.release()
-            data.append(buf)
-
-    else:
-        ### NOT WORKING ###
-        #TODO find a way to get videos directly into computer memory
-        for file in selected_files:
-            obj = s3_resource.Object(bucket_name, file)
-            response = obj.get()
-            result = response['Body'].read()
-            video = cv2.VideoCapture.open(BytesIO(result))
-            data.append(video)
+    # Delete local data unless specified
+    if(not bKeep_data):
+        shutil.rmtree(local_dir)
 
     return data
 
@@ -84,19 +95,32 @@ def retrieve_cam_details_from_database():
 
 
 def describe_s3_bucket():
+    """Plot the number of videos in the s3 bucket for each date.
+    Plot is saved locally under plots/01_exploratory.
+
+            Args:
+
+            Returns:
+
+            Raises:
+
+        """
+
+    # Set up boto3 session
     s3_session = boto3.Session(profile_name='dssg')
     bucket_name = 'air-pollution-uk'
     s3_resource = s3_session.resource('s3')
     my_bucket = s3_resource.Bucket(bucket_name)
 
+    # Get list of all dates in the s3 bucket
     objects = my_bucket.objects.filter(Prefix="raw/video_data_new/")
     files = [obj.key for obj in objects]
-
     dates = []
     for file in files:
         res = re.search("([0-9]{4}\-[0-9]{2}\-[0-9]{2})", file)
         dates.append(res.group())
 
+    # Plot a bar chart of the dates in the s3 bucket
     unique_dates, counts = np.unique(dates, return_counts=True)
     plt.figure()
     plt.bar(np.arange(unique_dates.shape[0]), counts)
